requirements.txt
openai == 0.28.1
deepdiff
ray
nltk
replicate
google-generativeai
